Development

+++ Production-Grade-Web-API/Domain/Entities/Cart.cs
@@ -3,7 +3,7 @@ public class Cart : BaseEntity
 {
     public Guid UserId { get; set; }
     public DateTime LastModifiedAt { get; set; } = DateTime.UtcNow;
-    public decimal CalculatedTotalPrice => CartItems.Sum(ci => ci.Quantity * ci.PriceAtAddTime);
+    public int CalculatedTotalPrice => (int)CartItems.Sum(ci => ci.Quantity * ci.PriceAtAddTime);
     public int TotalItemCount => CartItems.Sum(ci => ci.Quantity);
     public ICollection<CartItem> CartItems { get; set; } = new List<CartItem>();
     public ApplicationUser? User { get; set; }
